move embedded-hal_02 to subdir for serial
